+++ src/parser/hfbi/mod.rs
@@ -426,13 +426,6 @@ impl CampionamentoHFBI {
     {
         check_records_campionamento_hfbi::<T>(vec)
     }
-    #[deprecated(note = "v0.2 will drop visibility.\nConsider using check_records instead")]
-    pub fn check_record<T>(vec: Vec<T>) -> Result<Self, Vec<RecordCampionamentoHFBIError>>
-    where
-        T: RecordCampionamentoHFBI,
-    {
-        Self::check_records::<T>(vec)
-    }
 }
 
 impl AnagraficaHFBI {